ci: Source install dnsmasq-2.87
dnsmasq-2.86 shipped in Ubuntu jammy has a
known issue[1] which is fixed in dnsmasq-2.87
but it's not yet released with Ubuntu jammy.

Until fixed version is available in Ubuntu
jammy let's use source install instead of
using a older version from Ubuntu focal.

[1] https://lists.thekelleys.org.uk/pipermail/dnsmasq-discuss/2022q3/016562.html

Update from Julia:

Pushing forward the source fix again as ubuntu removed the
prior path we were using as a focal package and replaced
it with a package which is demonstrating the same basic issue.

Related-Bug: #2026757
Change-Id: I7ffcd167fc1e3a8c1192d766743bb5620d85ef35