Log when a periodic is completed
When debugigng the logs, it is sort of difficult to know when
actions complete for a specifc periodic task.

Example: Power sync, which includes an explicit sleep(0) to
yield control. The explicit sleep *is* important as it is
a low priority task. Where this quickly mixes things up
is when we're hunting database locking issues, and we
*must* verify that we *have exited* part of the code
completely.

Change-Id: Ie9eddedf6bca603845ff14e1ccbda665e3b9e5bd