Fix retry logic logging
logging has a method debug, and a static variable DEBUG, which are entirely
different. Tenacity requires the integer value which is passed in for logging
actions.

This is rooted in the examples published on tenacity logging, since the lower
level caller doesn't log in many cases, and our testing didn't catch this precise
case because we were only validating that logging was called, not that logging
worked. We also didn't see any of the errors related to CI resource
contention being lessened when the patch was running, and many of the chances
of lock conflicts being reduced in other fixes.

Duplicates the retry test *without* the logging mock, as not mocking the logging
would have yielded the break.

Change-Id: I4a65a044e90aff3cffae24f191e425bc75b5fb74