Add a list of children nodes unit test
Somehow, missed adding a test for this early on, while working on
validating the API client changes, I discovered that we seemed to
be missing unit tests for the method, because I couldn't get it
to work for the client, but then I realized I was just mixing
the nature of use up. C'est la vie!

Change-Id: I382ad3a011f3151abedb896f58c75e8b76b357fa