Log upon completion of power sync
Previously, we updated node_periodic so we understood from the
logs when a periodic task was completed, so we could understand
where things were at in our hunt for database lock racess.

In any event, we now explicitly log in the _sync_power_state
method of the conductor, because it is not a node_periodic.

Change-Id: Iaec9926fe031e65de4732ff0bc7988c5604d4755