Don't hold on to the API request dbapi
... And only execute the loader once.

We now use a cached dbapi instance, and then set the object
on the request to None, disconnecting the request from the
request from the database, which doesn't need to be held
open any longer than absolutely necessary.

Interestingly... locally, my total unit test time dropped with this
change from an aggregate around 200 seconds to 130-140 seconds. It
seems there was quite a bit of overhead in that call to begin with.

Change-Id: Ibffbc0b14d3b908f772b2cac8c6fedae432b267a