Task Manager: Log if the lock takes too long
In order to troubleshoot some of the locking issues, we need to
understand if we're getting held up on the start of the lock or
on the far end of the log upon completing the actual needful
work.

This change adds additional logging on the front end if it is
on the front end, and also sets the rpc object to None explicitly
before returning the update_node result, since it appears a stream
of them can cause issues for sqlite's release of the database from
one thread to another.

Change-Id: I8ef68ff62e8f93f7bc5bd9ee7c6ac8d96328f929