Log an exception from heartbeat
So it seems we can have weirdness here which
can get lost in the mix because heartbeat operations
get called with spawn_after on the task, and without
an error handler. The net effect seems to be we can
eat errors, which maybe we should log.

Logs a warning now and returns to the caller which will
exit the task.

Change-Id: I8d052e5d26396737bc7d807cbafdc317cfd2f21f