Change snmp job to not use a focal node
tl;dr, devstack no longer supports focal, and now errors which
results in the job failing.

Also changes the snmp job to utilize the test_ramdisk_iso
feature, as *opposed* to a full deployment because iPXE
shipped with Ubuntu no longer likes to chain boot in UEFI
mode to a block device. The easiest path, is just to run
a ramdisk in that case, which also sort of mirrors what
users of the SNMP power interface *tend* to do.

Related-Bug: #2034588

Change-Id: I276885b8f0492ead8cea38fe13826123131984ea