Handle Power On/Off for child node cases
So, there are cases where say you may have multiple DPUs in a
physical server, each card when fully operating can consume 100-150
watts. In some cases, these cards can have external power supplies,
but need the physical host in a running power-on state in order for
the device to be powered on.

Conversely, we also now power off the child nodes if the parent
node has been requested to be powered off, since we *really* don't
want cause inadvertent harm to the child node.

This is realistically a fix we should backport once we sort
through the details, if we agree this makes sense to do, as is.

Change-Id: Ib2bfe04cdaa82264ba8bb1e71477899bb6268179