Advanced vmedia deployment test ops
Adds an advanced operations standalone test which utilizes a new proposed
tempest test to execute against the API which exerises the dhcp-less
virtual media path AND passes the node through the rebuild scenario case
which has been identified as problematic in the past.

With this, we know it works, so \o/.

Presently as non-voting.

Change-Id: Ibb6f9228672966c3708227e37bead6a45648e177