eventlet monkey patch in unit tests earlier
It's important for consistent behavior to monkey_patch eventlet
before importing anything. While it makes an attempt to green any
existing created objects or locks, that code is buggy and fails
in some cases -- especially around rlocks.

It's not my belief that this resolves any specific bugs, but this
does reflect a better overall practice.

Change-Id: I57b2c91f9853287a08ee79ac87ae6e1767ddfb6f