Multiple CI updates/improvements
* Adds httpboot enabled CI

Sets the interfaces to be enabled on the standalone job
in anticipation for standalone job support for these
interfaces[0] and switches the boot from volume
job over to utilize httpboot by default.

* Phases out now-redundant ironic-standalone job

The ironic-standalone job does all the same work as the
ironic-standalone-redfish job, and IPMI CI jobs are covering
IPMI cases. So we don't need this anymore.

Also makes a note on another redfish job which is likely
redundant as well, although does not yet remove it.

* Fix IPMI Partition Job name

This is more of a test of OVN than of partition specifically,
in fact, it runs both wholedisk and partition jobs already. Make
the name more sensible.

[0]: https://review.opendev.org/c/openstack/ironic-tempest-plugin/+/902171

Co-Authored-By: Julia Kreger <juliaashleykreger@gmail.com>
Change-Id: I6c41b8f124e2e1fbc314243bf821153d79e2e09b