CI: Remove ironic-inspector-tempest-managed-non-standalone
The standalone job has not passed on the master branch in a
very long time. I stopped looking at the history as of August 2023,
but it was only passing on Xena and Wallaby stable branches.

A few things to note:

* It appears to be intended to exercise inspector as a WSGI app,
  and loads quite a bit of different configuration.
* Overall, it is failing due to the image it is trying to use
  (cirros) which doesn't support setting a bootloader up, but
  that is sort of required.
* Job is configured to netboot a node for an OS on the disk,
  but that is not something ironic does really anymore. Quite
  possibly, that is why the job is failing.
* Inspector, as a non-integrated thing, is on the path of
  deprecation at this point, so we don't really need to focus
  on this specific case anymore.

Overall, it seems like a job we should have just removed some
time ago. So... doing so.

Change-Id: I3aca63c183af863d9db1f27a4cfe0d6495bb03c2