Validate [deploy] image_server_auth_strategy
The [deploy] image_server_auth_strategy option effectively supports
only noauth or http-basic. However in case an operator set this to
an unsupported value, it results in incomplete behavior without any
explicit error. This is quite confusing and can cause difficulty in
troubleshooting.

This adds validation of the configuration to make operators aware of
any invalid value to improve operator's experience.

Change-Id: I2acf12f81bb64027eeb50924c2ac215c4270ca87