[stable-only] Fix tempest checkout/grenade run
Newer tempest plugins have been updated to default
towards system scope. Which is fine, but at the same
time it means our oldest branches are out of the standard
test window. So we need to pin the plugin to an older
version for the job to work as expected as it, by default,
tries to consume master branch unless specifically
over-ridden.

Also just disables the grenade run on this stable branch.
We're long past this being released, and neighboring branches
have moved to unmaintained. It just doesn't make sense
to keep executing.

Change-Id: I3b6d5a45cf6732b572a413458c15834eb3258970