packaging: reduce jsonschema revision
The VTEP switch support patch merged with a constraint of jsonschema
version 4.19 or above.

Except Debian only currently has 4.10, Centos 9 Stream only has 4.16,
and at present launchpad and the ubuntu mirror list is non-functional.

So in the interest of of packagers, we'll lower the version.

Note: I was able to successfully execute the unit tests with jsonschema
4.0.0 installed in the py3 virtualenv.

Change-Id: Ic3667a7663b7bd5dfad4665321d9c82cc08cc885