stable-only: pin proliantutils/scciclient to prevent break
Proliantutils 2.16.0 roughly times with the 2023.2
release of ironic and a switch to lextudio-pysnmp,
however in this branch of ironic, however this breaks
depending on order and collides with pysnmp namespace.

Also pins python-scciclient to <0.14.0 as to also not
pull in the dependency difference.

Also, Also, disables standalone, grenade, and metalsmith,
and snmp jobs from voting while we work to stabilize CI across
multiple branches.

Change-Id: Ibe3274d7fabfd4f06af8aba1af0957fa36e8d217