CI: Unpack cirros image partition jobs
Cirros, by default, as part of its initialization, copies the initial
ramdisk contents over the filesystem on disk. This changes the partition
image creation job so we do it upfront so the partition image looks like
and matches what we generally expect from a partition image as opposed
to just a kernel, ramdisk, and bootloader.

Change-Id: Idde30e33e9453f8564a7c3b9109c4e567146dee7