redfish: change default virtual media storage to local storage
This may be the most overridden default of Ironic, which means
we need to change the default value.

The default for ``[redfish]use_swift`` was historically ``true``,
however we've generally found that BMCs are particularlly sensitive
to extra characters in the URL as the characters may signify a
dynamically generated file, which would be problematic as virtual
media webservers also generally require range retrieval support.

This change makes the default ``false`` which should lead to one
less override for operators being necessary in practical operation.

Change-Id: Iad57b3c6423bced0e3cb6fb4e31aad6d805f26fa