[stable-only] wallaby: Reduce CI jobs
It looks like the move to unmaintained branching might have finally
done in the bulk of our tempest testing on this branch.

Some tests still work, but basically all tempest related jobs
are trying to pull in Xena stuffs without explicit config.

At this point, the value of the branch is to receieve critical
fixes being backported, and we've decided to drop the bulk of
the testing to match the status as we don't see fixing devstack
tempest invocation as critical to meet the branch's status.

Change-Id: Ia45d7bc8e033034c14061786b95b4f47f20268d0